@@ -1,5 +1,5 @@
 # This migration comes from forum (originally 20141012015557)
-class CreateForumPosts < ActiveRecord::Migration
+class CreateForumPosts < ActiveRecord::Migration[4.2]
   def change
     create_table :forum_posts do |t|
       t.string :title
